+++ src/main/java/lotto/model/Lotto.java
@@ -1,4 +1,4 @@
-package lotto.model.Lotto;
+package lotto.model;
 
 import java.util.Collections;
 import java.util.List;
